Adjusts function signatures.

+++ src/jetplot/colors.py
@@ -27,7 +27,7 @@ class Palette(list[ColorType]):
         """Return the palette as a Matplotlib colormap."""
         return LinearSegmentedColormap.from_list("", self)
 
-    def plot(self, figsize: tuple[float, float] = (5, 1)) -> tuple[Figure, NDArray[Axes]]:
+    def plot(self, figsize: tuple[int, int] = (5, 1)) -> tuple[Figure, NDArray[Axes]]:
         """Visualize the colors in the palette."""
         fig, axs = plt.subplots(1, len(self), figsize=figsize)
         for c, ax in zip(self, axs, strict=True): # pyrefly: ignore
